Move mixture random tests from test_distributions_random to test_mixture
